Basic idea of spectrum and LTAS implemented. Values were compared to the LTAS. Next steps:
Put the code in another class
Modify the tests and include a comparison with the current values
Refactor the code and run the tests :)
Create two or three aditional recordings and create tests for those.

Create all the entities and logic about it.
Start the investigation about the voice projection.